# internal/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
        },
        "description": "API for managing AIG Custom Platform AI security scanning tasks. Based on Tencent Zhuque Lab AI-Infra-Guard (https://github.com/Tencent/AI-Infra-Guard).",
        "title": "AIG Custom Platform 任务API",
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/platform/tasks": {
            "get": {
                "description": "Lists persisted platform tasks as TaskListResponse without polling the engine or mutating state. Users see only their own tasks; auditors and administrators see the stable global list. Items use the narrow TaskSummary only. page defaults to 1, page_size defaults to 20, values over 100 are capped at 100, and page cannot exceed 1000. Optional status and task_type filters accept only the documented exact canonical values and are applied before total and pagination.",
                "parameters": [
                    {
                        "default": 1,
                        "in": "query",
                        "maximum": 1000,
                        "minimum": 1,
                        "name": "page",
                        "type": "integer"
                    },
                    {
                        "default": 20,
                        "in": "query",
                        "maximum": 100,
                        "minimum": 1,
                        "name": "page_size",
                        "type": "integer"
                    },
                    {
                        "enum": [
                            "pending",
                            "dispatching",
                            "running",
                            "succeeded",
                            "failed",
                            "dispatch_failed",
                            "dispatch_unknown",
                            "cancelled"
                        ],
                        "in": "query",
                        "name": "status",
                        "type": "string"
                    },
                    {
                        "enum": [
                            "mcp_scan",
                            "ai_infra_scan",
                            "model_redteam_report",
                            "agent_scan"
                        ],
                        "in": "query",
                        "name": "task_type",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authorized stable task list.",
                        "schema": {
                            "$ref": "#/definitions/tasks.TaskListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page, page_size, status, or task_type; filter values must be exact canonical values."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Password change is incomplete or the Subject role is unsupported."
                    },
                    "500": {
                        "description": "Task query failed; internal details are not exposed."
                    }
                },
                "summary": "List authorized platform tasks",
                "tags": [
                    "platform-tasks"
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "description": "Breaking security hardening: create responses now use the browser-safe TaskDetail with bounded input_summary instead of the former internal task View; clients must migrate field access. The JSON body is limited to 256 KiB. Creates an owner-scoped platform task from the authenticated Cookie Subject. Idempotency-Key is required and unique per owner. A retry with the same persisted request payload returns the existing task without revalidating live governed references; reusing that key with a different payload returns 400 invalid task request. Only exact canonical task types are accepted and the platform maps them to private Agent aliases: mcp_scan params allow model_id and thread; ai_infra_scan params allow model_id and timeout; model_redteam_report params require model_id as a bounded string array plus eval_model_id and may include dataset numPrompts/randomSeed/promptColumn and techniques; agent_scan params require agent_id and eval_model_id. Every model_id/eval_model_id must resolve through the governed model boundary, and agent_id through the user or public Agent registry, before persistence; unknown or invisible references write no task. Unknown fields, nested credential objects, raw model credentials, and legacy aliases are rejected. Attachments are supplied only as at most ten opaque attachment IDs. Dispatch uses persisted claims and a global attempt budget. An uncertain network acknowledgement becomes dispatch_unknown and is never submitted automatically again; this is at-most-once dispatch, not exactly-once execution. Requires CSRF protection.",
                "parameters": [
                    {
                        "in": "header",
                        "maxLength": 128,
                        "name": "Idempotency-Key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "task",
                        "required": true,
                        "schema": {
                            "properties": {
                                "attachment_ids": {
                                    "items": {
                                        "type": "string",
                                        "maxLength": 128
                                    },
                                    "type": "array",
                                    "maxItems": 10
                                },
                                "content": {
                                    "type": "string",
                                    "maxLength": 32768
                                },
                                "country_iso_code": {
                                    "type": "string",
                                    "enum": [
                                        "",
                                        "zh",
                                        "zh_CN",
                                        "en"
                                    ]
                                },
                                "params": {
                                    "type": "object",
                                    "description": "Exact per-task schema described by this operation; arbitrary and credential-bearing fields are rejected."
                                },
                                "task_type": {
                                    "type": "string",
                                    "enum": [
                                        "mcp_scan",
                                        "ai_infra_scan",
                                        "model_redteam_report",
                                        "agent_scan"
                                    ]
                                }
                            },
                            "required": [
                                "task_type"
                            ],
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Browser-safe persisted task detail; raw request, attachment, identity, and engine fields are absent.",
                        "schema": {
                            "$ref": "#/definitions/tasks.TaskDetail"
                        }
                    },
                    "400": {
                        "description": "Fixed generalized body. Missing idempotency key, malformed input, or raw credentials return error=invalid task request; a missing or not ready attachment returns error=attachment unavailable. Identifiers, paths, states, content, and internal errors are never echoed.",
                        "schema": {
                            "$ref": "#/definitions/tasks.TaskCreateBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Role cannot create tasks, password change is incomplete, or CSRF is invalid."
                    },
                    "500": {
                        "description": "Persistence or durable audit failed; internal details are not exposed."
                    },
                    "503": {
                        "description": "Fixed safe error plus persisted browser-safe task detail; dispatch diagnostics are not exposed.",
                        "schema": {
                            "$ref": "#/definitions/tasks.TaskCreateErrorResponse"
                        }
                    }
                },
                "summary": "Create an idempotent platform task",
                "tags": [
                    "platform-tasks"
                ]
            }
        },
        "/api/v1/platform/tasks/{taskID}": {
            "get": {
                "description": "Short-poll browser-safe TaskDetail using the authenticated Cookie Subject. Users read only their own task; auditors and administrators have global read access. The input_summary is a narrow display projection; browser-supplied identity headers are ignored.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "taskID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Current platform task status.",
                        "schema": {
                            "$ref": "#/definitions/tasks.TaskDetail"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Password change is incomplete or the Subject role is unsupported."
                    },
                    "404": {
                        "description": "Task not found or not visible to this user."
                    },
                    "500": {
                        "description": "Task query failed; internal details are not exposed."
                    }
                },
                "summary": "Get owned task status",
                "tags": [
                    "platform-tasks"
                ]
            }
        },
        "/api/v1/platform/tasks/{taskID}/result": {
            "get": {
                "deprecated": true,
                "description": "Retired browser result route. After Cookie authentication and the password-change gate it always returns 410 Gone, regardless of task identity or state. It never reads or returns engine results. Authentication failure remains 401 and an incomplete password change remains 403.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "taskID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Password change is incomplete."
                    },
                    "410": {
                        "description": "Browser task result retrieval is permanently retired; use immutable reports."
                    }
                },
                "summary": "Retired browser task result route",
                "tags": [
                    "platform-tasks"
                ]
            }
        },
        "/api/v1/platform/tasks/{taskID}/cancel": {
            "post": {
                "description": "Cancels an owned task or, for administrators, any task. A task owned by another user is indistinguishable from an absent task and returns 404; auditors remain read-only with 403. Only the exact 204 response represents confirmed cancellation; clients may perform one status GET after an uncertain transport/server/non-contract-2xx outcome but must never automatically repeat the POST. Requires CSRF protection and records a durable audit event.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "taskID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task cancelled and audited."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Auditors or unsupported roles cannot cancel tasks; password change or CSRF may also be incomplete."
                    },
                    "404": {
                        "description": "Task not found or owned by another ordinary user."
                    },
                    "500": {
                        "description": "Cancellation or durable audit failed; internal details are not exposed."
                    }
                },
                "summary": "Cancel governed task",
                "tags": [
                    "platform-tasks"
                ]
            }
        },
        "/api/v1/platform/tasks/attachments": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "description": "Streams one private attachment owned by the authenticated user or administrator within the configured size limit. Before creation, one bounded batch (at most 100) of expired unbound uploading or ready records plus prior deleting tombstones is cleaned; records finalize only after private storage removal. Auditors cannot create attachments. Requires CSRF. The response exposes only an opaque attachment ID and safe metadata, never a storage name or path.",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "file",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Opaque private attachment metadata.",
                        "schema": {
                            "$ref": "#/definitions/tasks.AttachmentView"
                        }
                    },
                    "400": {
                        "description": "Invalid or oversized attachment."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Password change is incomplete or CSRF is invalid."
                    },
                    "413": {
                        "description": "Attachment exceeds the configured file-size limit."
                    },
                    "500": {
                        "description": "Storage, persistence, or durable audit failed; internal details are not exposed."
                    }
                },
                "summary": "Upload private task attachment",
                "tags": [
                    "platform-task-attachments"
                ]
            }
        },
        "/api/v1/platform/tasks/attachments/chunked": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "description": "Begins an owner-scoped chunked upload for an authenticated user or administrator after validating the declared total size. Before creation, one bounded batch (at most 100) of expired unbound uploading or ready records plus prior deleting tombstones is cleaned. Records are finalized only after private storage removal succeeds, so failures remain retryable. The configured TTL is 24 hours by default. Auditors cannot create attachments. Requires CSRF.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "attachment",
                        "required": true,
                        "schema": {
                            "properties": {
                                "filename": {
                                    "type": "string"
                                },
                                "size": {
                                    "format": "int64",
                                    "type": "integer"
                                }
                            },
                            "required": [
                                "filename",
                                "size"
                            ],
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Opaque uploading attachment metadata.",
                        "schema": {
                            "$ref": "#/definitions/tasks.AttachmentView"
                        }
                    },
                    "400": {
                        "description": "Invalid filename or size."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Password change is incomplete or CSRF is invalid."
                    },
                    "413": {
                        "description": "Declared size exceeds the configured file-size limit."
                    },
                    "500": {
                        "description": "Persistence or durable audit failed; internal details are not exposed."
                    }
                },
                "summary": "Begin private chunked upload",
                "tags": [
                    "platform-task-attachments"
                ]
            }
        },
        "/api/v1/platform/tasks/attachments/{attachmentID}/chunks": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "description": "Streams one non-empty bounded chunk. chunk_index must be non-negative and less than the configured ceil(max upload bytes / max chunk bytes); the default maximum is ten chunks. Owner users may write their own attachment; administrators may write any attachment; auditors are read-only. Cumulative size is enforced atomically. Requires CSRF.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "attachmentID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "chunk_index",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "formData",
                        "name": "chunk",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Chunk accepted."
                    },
                    "400": {
                        "description": "Invalid chunk index or attachment state."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Subject cannot write this attachment, password change is incomplete, or CSRF is invalid."
                    },
                    "404": {
                        "description": "Attachment not found."
                    },
                    "413": {
                        "description": "Chunk exceeds the configured chunk-size limit."
                    },
                    "500": {
                        "description": "Storage, persistence, or durable audit failed; internal details are not exposed."
                    }
                },
                "summary": "Upload private attachment chunk",
                "tags": [
                    "platform-task-attachments"
                ]
            }
        },
...
        "/api/v1/platform/tasks/attachments/{attachmentID}/download": {
            "get": {
                "description": "Downloads a ready or task-bound private attachment after authorization. An owner user may read their own attachment; other users receive 404 to avoid existence disclosure; auditors receive 403; administrators may read any attachment. Before an administrator cross-owner storage open, the server durably records a successful attachment.download_authorized event with sanitized authorization metadata; this records authorization, not stream delivery. Storage names and paths are never exposed.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "attachmentID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Streamed attachment.",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Attachment is not ready."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Auditor or unsupported role cannot download attachments; password change may also be incomplete."
                    },
                    "404": {
                        "description": "Attachment is absent, storage is absent, or it belongs to another ordinary user."
                    },
                    "500": {
                        "description": "Storage open or required authorization audit failed; no bytes or internal details are returned."
                    }
                },
                "summary": "Download authorized private attachment",
                "tags": [
                    "platform-task-attachments"
                ]
            }
        },
...
        "/api/v1/app/taskapi/tasks": {
            "post": {
                "deprecated": true,
                "description": "Retired browser execution endpoint. Requests receive 410 Gone only after the normal session, password-change, and CSRF checks; use POST /api/v1/platform/tasks with governed model IDs and opaque attachments.",
                "responses": {
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Password change is incomplete or CSRF is invalid."
                    },
                    "410": {
                        "description": "Endpoint retired; no task is submitted."
                    }
                },
                "summary": "Retired task creation endpoint",
                "tags": [
                    "taskapi"
                ]
            }
        },
...
    },
    "definitions": {
...
        "tasks.TaskCreateBadRequestResponse": {
            "description": "Fixed generalized bad-request body; it never echoes attachment identifiers, paths, states, request content, or internal errors.",
            "properties": {
                "error": {
                    "enum": [
                        "invalid task request",
                        "attachment unavailable"
                    ],
                    "type": "string"
                }
            },
            "required": [
                "error"
            ],
            "type": "object"
        },
...
        "tasks.TaskListResponse": {
            "properties": {
                "items": {
                    "items": {
                        "$ref": "#/definitions/tasks.TaskSummary"
                    },
                    "type": "array"
                },
                "page": {
                    "minimum": 1,
                    "type": "integer"
                },
                "page_size": {
                    "maximum": 100,
                    "minimum": 1,
                    "type": "integer"
                },
                "total": {
                    "format": "int64",
                    "minimum": 0,
                    "type": "integer"
                }
            },
            "required": [
                "items",
                "total",
                "page",
                "page_size"
            ],
            "type": "object"
        },
...
    }
}